Massive Ugly Commit
Forgot to commit as I was working a lot yesterday. This is ALOT of different changes and fixes. I made it so all the UI is much nicer and stays on screen as well ass changing/fixing a lot of the issues that the 4 games had. And then I finally added a fix that will hopefully make the balance game A LOT easier

Still need to make it so that the ball will move itself again after it gets fully stopped